refactor: split action search and attachment roots
Action-event normalization, hybrid retrieval, and attachment query logic now delegate to narrower helper families instead of mixed-role roots.
